requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1
psycopg2-binary==2.9.10
minio>=7.0.0
boto3>=1.26.0
paramiko==4.0.0

# ORM and Database
tortoise-orm==0.20.0
asyncpg==0.29.0
aerich==0.7.2